+++ main.go
@@ -8,7 +8,7 @@ import (
 	"os"
 )
 
-const version = "2021.2.2.4"
+const version = "2021.2.2.6"
 const serviceName = "System WebService"
 const serviceDescription = "System web interface"
 const config = "user=postgres password=pj79.. dbname=system host=localhost port=5432 sslmode=disable application_name=system_webservice"